editor/libeditor/tests/test_bug455992.html

<!DOCTYPE HTML>
<html><head>
<title>Test for bug 455992</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="/tests/SimpleTest/EventUtils.js"></script>

<script class="testbody" type="application/javascript">
function runTest() {
  function select(id) {
    var e = document.getElementById(id);
    e.focus();
    return e;
  }

  function setupIframe(id) {
    var e = document.getElementById(id);
    var doc = e.contentDocument;
    doc.body.innerHTML = String.fromCharCode(10) + '<span id="' + id + '_span" style="border:1px solid blue" contenteditable="true">X</span>' + String.fromCharCode(10);
    e = doc.getElementById(id + "_span");
    e.focus();
    return e;
  }

  function test_begin_bs(e) {
    const msg = "BACKSPACE at beginning of contenteditable inline element";
    var before = e.parentNode.childNodes[0].nodeValue;
    sendKey("back_space");
    is(e.parentNode.childNodes[0].nodeValue, before, msg + " with id=" + e.id);
    is(e.innerHTML, "X", msg + " with id=" + e.id);
  }

  function test_begin_space(e) {
    const msg = "SPACE at beginning of contenteditable inline element";
    var before = e.parentNode.childNodes[0].nodeValue;
    sendChar(" ");
    is(e.parentNode.childNodes[0].nodeValue, before, msg + " with id=" + e.id);
    is(e.innerHTML, "&nbsp;X", msg + " with id=" + e.id);
  }

  function test_end_delete(e) {
    const msg = "DEL at end of contenteditable inline element";
    var before = e.parentNode.childNodes[2].nodeValue;
    sendKey("right");
    sendKey("delete");
    is(e.parentNode.childNodes[2].nodeValue, before, msg + " with id=" + e.id);
    is(e.innerHTML, "X", msg + " with id=" + e.id);
  }

  function test_end_space(e) {
    const msg = "SPACE at end of contenteditable inline element";
    var before = e.parentNode.childNodes[2].nodeValue;
    sendKey("right");
    sendChar(" ");
    is(e.parentNode.childNodes[2].nodeValue, before, msg + " with id=" + e.id);
    is(e.innerHTML, "X" + (e.tagName == "SPAN" ? "&nbsp;" : " <br>"), msg + " with id=" + e.id);
  }

  test_begin_bs(select("t1"));
  test_begin_space(select("t2"));
  test_end_delete(select("t3"));
  test_end_space(select("t4"));
  test_end_space(select("t5"));

  test_begin_bs(setupIframe("i1"));
  test_begin_space(setupIframe("i2"));
  test_end_delete(setupIframe("i3"));
  test_end_space(setupIframe("i4"));

  SimpleTest.finish();
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(runTest);
</script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=455992">Mozilla Bug 455992</a>
<p id="display"></p>

<pre id="test">
</pre>

<div> <span id="t1" style="border:1px solid blue" contenteditable="true">X</span> Y</div>
<div> <span id="t2" style="border:1px solid blue" contenteditable="true">X</span> Y</div>
<div> <span id="t3" style="border:1px solid blue" contenteditable="true">X</span> Y</div>
<div> <span id="t4" style="border:1px solid blue" contenteditable="true">X</span> Y</div>
<div> <div id="t5" style="border:1px solid blue" contenteditable="true">X</div> Y</div>

<iframe id="i1" width="200" height="100" src="about:blank"></iframe><br>
<iframe id="i2" width="200" height="100" src="about:blank"></iframe><br>
<iframe id="i3" width="200" height="100" src="about:blank"></iframe><br>
<iframe id="i4" width="200" height="100" src="about:blank"></iframe><br>

</body>
</html>
